TraitWarped this is a mild mess, but it is functional enough.
adds Thaumcraft warp to a tool by manipulating the "TC.WARP" tag, which recalculates when the tool is modified.
all levels use the same localization keys, and also slap "potion.potency.LEVEL" (where "LEVEL" is the trait instance level) onto the end of it, for easy extensibility.

also has a mildly cursed system for fetching it, where its value in TinkerTantrumTraits is a map that should be accessed from TraitWarped#getInstance, ideally before the game is fully loaded.